taunt animation
g2mApose is used to get a g2 into mixamo an animation is aplied to this and exported as fbx. To import into daz
g2mTauntPP is a pose preset created from the imported mixamo back into daz this can be applied to a default G8 and then corrected with the g8mMixamoAnimationFixer script.
PPg8mtaunt is a pose present of G8 after the mixamo g8mMixamoAnimationFixer script is used. PPg8mtaunt can be applied to any G8 figure.
man1taunt is a full scene manga shaders and G8 taunt